+++ src/main/scala/com/wavesplatform/atomicswap/Playground.scala
@@ -34,11 +34,11 @@ object Playground extends App {
   // 3N2Bk9EtqW5PGuc38pYaEUKzqsMjEydwXFi
   private val wavesUser =
     PrivateKeyAccount.fromSeed(Base58.encode(
-      "sad capable gospel wage bean evoke hundred crawl logic question cheese outer leader author decrease".getBytes), 1, p.wavesNetwork)
+      "sad capable gospel wage bean evoke hundred crawl logic question cheese outer leader author decrease".getBytes), 1, p.wavesNetwork.toByte)
   private val wavesUserTmpPrivateKey =
     PrivateKeyAccount.fromSeed(Base58.encode(
-      "sad capable gospel wage bean evoke hundred crawl logic question cheese outer leader author decreassrr34gg4".getBytes), 1, p.wavesNetwork)
-  private val wavesUserTmpPublicKey = new PublicKeyAccount(wavesUserTmpPrivateKey.getPublicKey, p.wavesNetwork)
+      "sad capable gospel wage bean evoke hundred crawl logic question cheese outer leader author decreassrr34gg4".getBytes), 1, p.wavesNetwork.toByte)
+  private val wavesUserTmpPublicKey = new PublicKeyAccount(wavesUserTmpPrivateKey.getPublicKey, p.wavesNetwork.toByte)
   // mzqD1A9pAJyELfMAWUrqcz5K8WfQEcXTPY
   private val wavesUserBitcoinECKey = ECKey.fromPrivate(KeysUtil.privateKeyBytesFromWIF(
     "91jVRNUJWu9ousWk6LdeUFRPcdFDmBiw5jnpYLogE2Ki8AwiZQg"))
@@ -47,13 +47,13 @@ object Playground extends App {
   private val bitcoinUserBitcoinECKey = ECKey.fromPrivate(KeysUtil.privateKeyBytesFromWIF(
     "91h8oWwuCkzxs979qFNXLF9raNxewMYozW2MnT9pPJ8mids26Wi"))
   private val bitcoinUserWavesAccount = PrivateKeyAccount.fromSeed(Base58.encode(
-    "sad capable gospel wage bean evoke hundred crawl logic question cheese outer leader author trololo!".getBytes), 1, p.wavesNetwork)
+    "sad capable gospel wage bean evoke hundred crawl logic question cheese outer leader author trololo!".getBytes), 1, p.wavesNetwork.toByte)
 
   // todo many outputs support
   private val bitcoinUserBitcoinOutInfo = BitcoinInputInfo("04891274037778941a771bcaad74569a8f5ca53b447acbbd711f5c7b25d70a1c", 0,
     ScriptBuilder.createOutputScript(bitcoinUserBitcoinECKey.toAddress(p.bitcoinNetworkParams)), bitcoinUserBitcoinECKey.getPrivKeyBytes)
 
-  // TX1 - Alice Waves -> scr1 money to tmp account + TX0-1 fee
+  // TX1 - Alice Waves -> scr1 money to tmp account + TX1-1 fee  + TX3 fee
   // TX1-1 - scr1 set script to tmp account
 
   val n = new Node("https://testnode1.wavesnodes.com/")
@@ -65,14 +65,14 @@ object Playground extends App {
 
   val tx2 = BitcoinSide.createAtomicSwapTransaction(bitcoinUserBitcoinOutInfo, bitcoinUserBitcoinECKey.getPubKey, wavesUserBitcoinECKey.getPubKey, p.startTimestamp.toSeconds + p.minutesTimeout.toSeconds)
 
-  // TX3 - Service [normal case] - scr1 -> Bob Waves address
-  // TX4 - Service [normal case] - scr1 -> Alice Bitcoin address
+  // TX3 - Service (or someone) [normal case] - scr1 -> Bob Waves address
+  // TX4 - Alice [normal case] - scr1 -> Alice Bitcoin address
 
   val tx3 = ServiceSide.accomplishBitcoinSwapTransaction(tx2.id, tx2.underlying.getOutput(0).getScriptPubKey, serviceX, wavesUserBitcoinECKey.getPrivKeyBytes, wavesUserBitcoinECKey.getPubKey)
   val tx4 = ServiceSide.accomplishWavesSwapTransaction(wavesUserTmpPublicKey, bitcoinUserWavesAccount, serviceX, p.wavesSmartFee)
 
   // TX5 - Service (or someone) [failed case] - scr1 (TX0-1) -> Alice Waves address
-  // TX6 - Service (or someone) [failed case] - scr2 (TX1) -> Bob Bitcoin address
+  // TX6 - Bob [failed case] - scr2 (TX1) -> Bob Bitcoin address
 
   val tx5 = ServiceSide.recoverBitcoinSwapTransaction(tx2.id, tx2.underlying.getOutput(0).getScriptPubKey, p.bitcoinAmount.minus(p.bitcoinFee.multiply(2)), bitcoinUserBitcoinECKey.getPrivKeyBytes, bitcoinUserBitcoinECKey.getPubKey)
   val tx6 = ServiceSide.recoverWavesSwapTransaction(wavesUserTmpPublicKey, wavesUser, p.wavesSmartFee)
